requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.1
python-dotenv==1.0.0

# Deep Learning
torch==2.1.1
torchvision==0.16.1
efficientnet-pytorch==0.7.1
captum==0.6.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.21.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Database
pymongo==4.6.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Visualization
matplotlib==3.7.1
seaborn==0.13.0
plotly==5.18.0

# Documentation and Reporting
reportlab==4.0.7
openpyxl==3.1.2
python-jose==3.3.0

# Development Tools
black==23.11.0
isort==5.12.0
mypy==1.7.0
flake8==6.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Progress Bars and Logging
tqdm==4.66.1
loguru==0.7.2
